requirements.txt
flask
flask-socketio
python-docx
pandas
openpyxl
eventlet
matplotlib
